+++ crates/ruma-events/CHANGELOG.md
@@ -1,5 +1,14 @@
 # [unreleased]
 
+Breaking changes:
+
+* `room::name::NameEventContent` now uses a custom `RoomName` type for its
+  `name` field and makes it public, in response the constructor and `name`
+  accessor had their types updated too
+* Replace `InvalidEvent` by a more specific `FromStringError` for room name
+  validation
+* Remove unused `FromStrError`
+
 # 0.23.3
 
 Improvements: